requirements.txt
playwright
python-dotenv
loguru
tenacity
pytest
pytest-asyncio